fix(justerm-renderer)!: the cluster-index column is u32, matching what core emits (#627)
Slice A of #627. `justerm-core` widened the frame's grapheme-cluster index to `u32`
in #621 — the table holds one entry per combining cell of the viewport, and the frame
header stores `cols` and `rows` as `u16` **each**, so a legal viewport holds far more
cells than a `u16` can number. This crate stayed narrow, and `justerm-web` bridged the
two by truncating.

Measured on the JS side while filing #627: index 65536 becomes `0`, which reads as
"no cluster" and silently drops the marks; 65537 becomes `1`, which resolves **the
wrong cluster**. No error at any layer. Separately, the coercion turned a zero-copy
pass-through into an unconditional per-frame copy of the whole column — that half is
not conditional on any large input and is the reason this is worth fixing at a
reachability the truncation alone would not justify.

`DamageFrame::extra` and `apply_damage`'s parameter are the only two declarations; the
read site needed no change (it already widened to `usize`).

## Test-trust

`a_cluster_index_above_u16_resolves_to_its_own_cluster` builds a side table whose
*last* entry is the one the index names and whose *first* is what a truncating chain
would land on, so the two are distinguishable. Strict RED was unavailable — with the
column still `&[u16]` the test does not compile, the "compilation enforces
completeness" case — so discrimination was proven by mutation instead: re-narrowing
the read to `as u16 as usize` reddens exactly this test, and on the real symptom
(`e◌́` where `e◌̈` was named) rather than on a bare inequality. Mutation verified
present in the file before its result was trusted (#577).

## Demo pages

Seven `apply_damage` call sites across six pages passed `extra` as a `Uint16Array`.
Widened by position — the parameter sits immediately before `side_table`, which is
what makes it identifiable in calls with several same-typed arguments. These are not
decoration: `test:proofs` drives them in a real browser, so a wrong slot would have
failed at the wasm boundary rather than silently.

## Gates

Renderer is outside every cargo umbrella, so each was named explicitly:
`cargo test/fmt --check/clippy --target wasm32/build --target wasm32/doc -D warnings`
`--manifest-path justerm-renderer/Cargo.toml` — all exit 0. `pnpm run test:unit` exit 0.
`pnpm run test:proofs` **90 passed**, on `wasm-pack 0.15.0` — verified against
`WASM_PACK_VERSION` in the workflow first, since a different local version would have
proven pixels against different codegen than CI judges (#616).

Step 6 is **N/A**: grepped `justerm-renderer/README.md` and the `frame-adapter`
territory for the column and its width — no prose describes either, so nothing
describes the old behaviour.

## Not in this slice

`justerm-web` still coerces with `asU16(frame.extra)` and still declares
`RendererBackend.extra: Uint16Array`. It consumes the **published** renderer at
`^0.8.0`, which under npm's 0.x caret cannot resolve `0.9.0`, so the web half cannot
land until this ships on a `renderer-v*` tag. Sequencing is on #627; the release plan
is #633.

+++ justerm-renderer/src/webgl.rs
@@ -1809,7 +1809,7 @@ impl JustermRenderer {
         fg: &[u32],
         bg: &[u32],
         flags: &[u16],
-        extra: &[u16],
+        extra: &[u32],
         side_table: Vec<String>,
         // #520: the span-ordered underline colour column (SGR 58), tagged-u32 like `fg`/`bg`.
         // Optional + TRAILING for the same reason as `apply_frame` — a caller that predates it
